Refactor `spawn_obstacles`
Your `#spawn_obstacles` method wasn't quite honest about what it was
doing; it was actually more of a general timer handling method. I
changed this to a `#handle_timer` method and extracted the object
spawning logic to an actual `#spawn_obstacles` method.

Because of this more general implementation, I was able to use it in the
constructor as well, rather than repeating the same hashes twice.